Add exec photos
